AMIGOS/requirements.txt
dataclasses==0.6
future==0.18.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
six==1.15.0
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
xlrd==1.2.0
